Add detect body forward vector
Added the ability to attempt to detect the users avatar body forward
vector. As there is no torso sensor this can only be at best an educated
guess using the headset and controllers. Caches for the headset position,
orientation, and both controllers positions are stored and updated every
frame. Each cache is then averaged and the current readings are compared
with it. The comparison will be used to detect how much the user is
moving. Body forward vector will only be updated when either the hands
and headset are static or the headset is static and the
headset/controller vectors align. The headset vector is a rotation
vector pointing in the direction the headset is facing. The controllers
vector is the perpendicular vector coming from the line between the two
controller. yada yada yada The forward vector is determined.