Finalize boneIK
Finalized the boneIK to function reasonably well. The pole target was
split to two seperate ones for both controllers. The pole position
changes based on the controllers in a "micro" and "macro" positions.
The micro positions are between the shoulders. The macro positions
are outside of the shoulders. There is noticable shift when crossing
the threshold but it is not unreasonable, and this way leads to much
more accurate arm positioning.